♻️ Refactor: constant, types 파일 옮기기
