+++ src/app/test/joy/page.tsx
@@ -2,38 +2,31 @@ import { Suspense } from 'react';
 
 import { SearchParamsType } from '@common/types/routes.types';
 
-import parseEnumQueryParam from '@common/utils/parseEnumQueryParam';
+import parseEnum from '@common/utils/parse-enum';
 
-import { LeftTabType } from '@features/ai-meeting-manager/constants/tabs';
-import { SnsFileTabType } from '@features/sns-event-assistant/constants/tabs';
-import {
-  SurveyQuestionTabType,
-  TeamMoodReportTabType,
-} from '@features/team-mood-tracker/constants/tabs';
-
+import LeftTab from '@features/ai-meeting-manager/components/LeftTab/LeftTab.client';
+import { LeftTabType } from '@features/ai-meeting-manager/components/LeftTab/LeftTab.types';
 import LeftPanel from '@features/ai-meeting-manager/components/panels/LeftPanel/LeftPanel.server';
 import RightPanel from '@features/ai-meeting-manager/components/panels/RightPanel/RightPanel.server';
-import LeftTab from '@features/ai-meeting-manager/components/tabs/LeftTab/LeftTab.client';
 import SnsFilePanel from '@features/sns-event-assistant/components/SnsFilePanel/SnsFilePanel.server';
 import SnsFileTab from '@features/sns-event-assistant/components/SnsFileTab/SnsFileTab.client';
+import { SnsFileTabType } from '@features/sns-event-assistant/components/SnsFileTab/SnsFileTab.types';
 
 import SurveyQuestionPanel from '@features/team-mood-tracker/componets/panels/SurveyQuestionPanel/SurveyQuestionPanel.server';
 import TeamMoodReportPanel from '@features/team-mood-tracker/componets/panels/TeamMoodReportPanel/TeamMoodReportPanel.server';
 import SurveyQuestionTab from '@features/team-mood-tracker/componets/tabs/SurveyQuestionTab/SurveyQuestionTab.client';
+import { SurveyQuestionTabType } from '@features/team-mood-tracker/componets/tabs/SurveyQuestionTab/SurveyQuestionTab.types';
 import TeamMoodReportTab from '@features/team-mood-tracker/componets/tabs/TeamMoodReportTab/TeamMoodReportTab.client';
+import { TeamMoodReportTabType } from '@features/team-mood-tracker/componets/tabs/TeamMoodReportTab/TeamMoodReportTab.types';
 
 const TestPage = async ({ searchParams }: { searchParams: Promise<SearchParamsType> }) => {
   // // TAB_AI 회의 진행 매니저_좌측
   // const { leftTab } = await searchParams;
-  // const formattedLeftTab = parseEnumQueryParam(leftTab, LeftTabType, LeftTabType.MEETING_SUMMARY);
+  // const formattedLeftTab = parseEnum(leftTab, LeftTabType, LeftTabType.MEETING_SUMMARY);
 
   // // TAB_SNS 이벤트 어시스턴트
   // const { snsFileTab } = await searchParams;
-  // const formattedTab = parseEnumQueryParam(
-  //   snsFileTab,
-  //   SnsFileTabType,
-  //   SnsFileTabType.PARTICIPANT_LIST,
-  // );
+  // const formattedTab = parseEnum(snsFileTab, SnsFileTabType, SnsFileTabType.PARTICIPANT_LIST);
   // // 실제로는 서버에서 count 가져오기
   // const participantCount = 10;
   // const winnerCount = 10;
@@ -45,7 +38,7 @@ const TestPage = async ({ searchParams }: { searchParams: Promise<SearchParamsTy
   };
 
   const { moodTab } = await searchParams;
-  const formattedTab = parseEnumQueryParam(
+  const formattedTab = parseEnum(
     moodTab,
     TeamMoodReportTabType,
     TeamMoodReportTabType.TEAM_MOOD_REPORT,
