+++ src/features/ai-meeting-manager/components/LeftTab/LeftTab.constants.ts
@@ -1,0 +1,6 @@
+import { LeftTabType } from './LeftTab.types';
+
+export const LeftTabLabels: Record<LeftTabType, string> = {
+  [LeftTabType.MEETING_SUMMARY]: 'AI 회의록',
+  [LeftTabType.MEETING_VOICE_LOG]: '음성 기록',
+};
