+++ src/features/ai-meeting-manager/components/LeftTab/LeftTab.types.ts
@@ -1,0 +1,8 @@
+export interface LeftTabProps {
+  current: LeftTabType;
+}
+
+export enum LeftTabType {
+  MEETING_SUMMARY = 'MEETING_SUMMARY',
+  MEETING_VOICE_LOG = 'MEETING_VOICE_LOG',
+}
