+++ src/features/ai-meeting-manager/components/panels/RightPanel/RightPanel.constants.ts
@@ -1,0 +1,6 @@
+import { RightTabType } from './RightPanel.types';
+
+export const RightTabLabels: Record<RightTabType, string> = {
+  [RightTabType.AI_QUESTIONS]: 'AI 질문 기록',
+  [RightTabType.AI_RECOMMENDATIONS]: 'AI 추천 질문',
+};
