+++ src/features/ai-meeting-manager/components/panels/RightPanel/RightPanel.types.ts
@@ -1,0 +1,5 @@
+// 실제 탭 역할은 안 함, ui용
+export enum RightTabType {
+  AI_QUESTIONS = 'AI_QUESTIONS',
+  AI_RECOMMENDATIONS = 'AI_RECOMMENDATIONS',
+}
