+++ src/features/sns-event-assistant/components/SnsFilePanel/SnsFilePanel.types.ts
@@ -1,4 +1,4 @@
-import { SnsFileTabType } from '@features/sns-event-assistant/constants/tabs';
+import { SnsFileTabType } from '../SnsFileTab/SnsFileTab.types';
 
 export interface SnsFilePanelProps {
   tab: SnsFileTabType;
