+++ src/features/sns-event-assistant/components/SnsFileTab/SnsFile.constants.ts
@@ -1,8 +1,4 @@
-export enum SnsFileTabType {
-  PARTICIPANT_LIST = 'PARTICIPANT_LIST',
-  WINNER_LIST = 'WINNER_LIST',
-  SNS_LINK = 'SNS_LINK',
-}
+import { SnsFileTabType } from './SnsFileTab.types';
 
 export const SnsFileTabLabels: Record<SnsFileTabType, string> = {
   [SnsFileTabType.PARTICIPANT_LIST]: '참여자 리스트',
