+++ src/features/sns-event-assistant/components/SnsFileTab/SnsFileTab.types.ts
@@ -1,6 +1,10 @@
-import { SnsFileTabType } from '@features/sns-event-assistant/constants/tabs';
-
 export interface TabProps {
   current: SnsFileTabType;
   counts: Partial<Record<SnsFileTabType, number>>;
+}
+
+export enum SnsFileTabType {
+  PARTICIPANT_LIST = 'PARTICIPANT_LIST',
+  WINNER_LIST = 'WINNER_LIST',
+  SNS_LINK = 'SNS_LINK',
 }
