+++ src/features/team-mood-tracker/componets/tabs/SurveyQuestionTab/SurveyQuestionTab.constants.ts
@@ -1,0 +1,6 @@
+import { SurveyQuestionTabType } from './SurveyQuestionTab.types';
+
+export const SurveyQuestionTabLabels: Record<SurveyQuestionTabType, string> = {
+  [SurveyQuestionTabType.SURVEY_GENERATE]: '설문 문항 생성',
+  [SurveyQuestionTabType.SURVEY_LIST]: '설문 문항',
+};
