+++ src/features/team-mood-tracker/componets/tabs/TeamMoodReportTab/TeamMoodReportTab.constants.ts
@@ -1,0 +1,7 @@
+import { TeamMoodReportTabType } from './TeamMoodReportTab.types';
+
+export const TeamMoodReportTabLabels: Record<TeamMoodReportTabType, string> = {
+  [TeamMoodReportTabType.TEAM_MOOD_REPORT]: '팀 분위기 리포트',
+  [TeamMoodReportTabType.ANSWER_SUMMARY]: '응답',
+  [TeamMoodReportTabType.SURVEY_LIST]: '설문 문항',
+};
