+++ src/features/team-mood-tracker/componets/tabs/TeamMoodReportTab/TeamMoodReportTab.types.ts
@@ -1,6 +1,10 @@
-import { TeamMoodReportTabType } from '@features/team-mood-tracker/constants/tabs';
-
 export interface TeamMoodReportTabProps {
   current: TeamMoodReportTabType;
   counts: Partial<Record<TeamMoodReportTabType, number>>;
+}
+
+export enum TeamMoodReportTabType {
+  TEAM_MOOD_REPORT = 'TEAM_MOOD_REPORT',
+  ANSWER_SUMMARY = 'ANSWER_SUMMARY',
+  SURVEY_LIST = 'SURVEY_LIST',
 }
